==> Scripts/00-packages.R <==
#commonly used packages
devtools::install_github("KluaneRedSquirrelProject/krsp")
library(krsp)
library(devtools)
library(data.table)
library(dplyr)
library(lubridate)
library(sf)
library(ggplot2)
library(ggmap)
library(hms)
library(sp)
library(tidyverse)
library(tidyr)
library(concaveman)
library(adehabitatHR)
library(amt)
library(raster)
library(arm)
library(rsq)
library(stringr)
library(effsize)
library(stats)
library(ggeffects)
library(MASS)
library(vcd)
library(gamlss)
library(gamlss.dist)
library(quantreg)
library(mgcv)
library(jtools)
library(pROC)
library(interactions)
library(ggforce)
library(dunn.test)
library(effectsize)
library(lme4)
library(influence.ME)
library(Hmisc)
library(lmerTest)
library(emmeans)
library(merTools)
library(glmmTMB)
library(DHARMa)
